+++ packages/cosmwasm/src/types.ts
@@ -22,6 +22,6 @@ export function parseWasmData({ key, val }: WasmData): Model {
 
 /**
  * An object containing a parsed JSON document. The result of JSON.parse().
- * This doen't privide any type safety over `any` but expresses intent in the code.
+ * This doesn't provide any type safety over `any` but expresses intent in the code.
  */
 export type JsonObject = any;