+++ packages/cosmwasm/types/index.d.ts
@@ -1,3 +1,4 @@
+export { isValidBuilder } from "./builder";
 export { Expiration } from "./interfaces";
 export { setupWasmExtension, WasmExtension } from "./lcdapi/wasm";
 export { BankMsg, CosmosMsg, CustomMsg, StakingMsg, WasmMsg } from "./cosmosmsg";
@@ -31,6 +32,7 @@ export {
   VotersResult,
 } from "./cw3cosmwasmclient";
 export {
+  ChangeAdminResult,
   ExecuteResult,
   CosmWasmFeeTable,
   InstantiateOptions,
@@ -54,3 +56,4 @@ export {
   MsgUpdateAdmin,
   MsgStoreCode,
 } from "./msgs";
+export { JsonObject, parseWasmData, WasmData } from "./types";